+++ Tree/strategy_filling.py
@@ -36,7 +36,7 @@ class StrategyFilling():
 		# filling strategy
 		# we will fill strategy with an uniform probability, but it has to be
 		# zero for hands that are not possible on corresponding board
-		node.strategy = np.zeros([len(node.children), CC], dtype=np.float32)
+		node.strategy = np.zeros([len(node.children), CC], dtype=arguments.dtype)
 		# setting probability of impossible hands to 0
 		for i in range(len(node.children)):
 			child_node = node.children[i]
@@ -55,7 +55,7 @@ class StrategyFilling():
 		if node.terminal:
 			return
 		value = 1.0 / len(node.children)
-		node.strategy = np.full([len(node.children), CC], value, dtype=np.float32)
+		node.strategy = np.full([len(node.children), CC], value, dtype=arguments.dtype)
 
 
 	def _fill_uniform_dfs(self, node):
